requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
dataclasses-json==0.6.7
Django==4.2.10
djangorestframework==3.14.0
drf-yasg==1.21.10
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
inflection==0.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_pydantic==0.6
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.59
langchain-ollama==0.3.2
langchain-text-splitters==0.3.8
langsmith==0.3.42
marshmallow==3.26.1
mcp==1.8.0
mcp-use==1.2.12
multidict==6.4.3
mypy_extensions==1.1.0
numpy==2.2.5
ollama==0.4.8
orjson==3.10.18
packaging==24.2
Pillow==10.1.0
propcache==0.3.1
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.2
sse-starlette==2.3.4
starlette==0.46.2
tenacity==9.1.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
yarl==1.20.0
zstandard==0.23.0
